video:fbdev:sirfsoc: Clean compile warning
drivers/video/fbdev/sirfsoc/g2d/g2d.c:1216:11: warning: passing argument 1 of 'dev_err' from incompatible pointer type
drivers/video/fbdev/sirfsoc/vdss/layer_screen.c:916:16: warning: unused variable 'flags' [-Wunused-variable]
drivers/video/fbdev/sirfsoc/vdss/layer_screen.c:913:26: warning: unused variable 'ldata' [-Wunused-variable]
drivers/video/fbdev/sirfsoc/vdss/vpp.c:1367:3: warning: enumeration value 'VDSS_FIELD_NONE' not handled in switch [-Wswitch]
drivers/video/fbdev/sirfsoc/vdss/vpp.c:1367:3: warning: enumeration value 'VDSS_FRAME_TOP' not handled in switch [-Wswitch]
drivers/video/fbdev/sirfsoc/vdss/vpp.c:1367:3: warning: enumeration value 'VDSS_FRAME_BOTTOM' not handled in switch [-Wswitch]
drivers/video/fbdev/sirfsoc/vdss/vpp.c:1420:22: warning: suggest parentheses around comparison in operand of '|' [-Wparentheses]
drivers/video/fbdev/sirfsoc/vdss/vpp.c:1567:23: warning: passing argument 1 of '__vpp_seq_type' from incompatible pointer type
drivers/video/fbdev/sirfsoc/vdss/vpp.c:1590:31: warning: passing argument 3 of '__vpp_set_ctrl' from incompatible pointer type
drivers/video/fbdev/sirfsoc/vdss/screen-sysfs.c:118:19: warning: initialization from incompatible pointer type
drivers/video/fbdev/sirfsoc/vdss/screen-sysfs.c:118:19: warning: (near initialization for 'screen_ktype.default_attrs')
drivers/video/fbdev/sirfsoc/vdss/dcu.c:173:6: warning: unused variable 'dst_skip' [-Wunused-variable]
drivers/video/fbdev/sirfsoc/vdss/dcu.c:172:6: warning: unused variable 'src_skip' [-Wunused-variable]
drivers/video/fbdev/sirfsoc/vdsscomp/vdsscomp.c:245:33: warning: comparison between 'enum vdss_pixelformat' and 'enum vdsscomp_pixelformat' [-Wenum-compare]

+++ drivers/video/fbdev/sirfsoc/vdsscomp/vdsscomp.c
@@ -242,8 +242,8 @@ int vdsscomp_gralloc_queue(struct vdsscomp_setup_data *d,
 			/* If the fmt is changed, disable the layer */
 			if (disp->layers[layer].enabled) {
 				l->layer->get_info(l->layer, &layer_info);
-				if (layer_info.src_surf.fmt !=
-				    disp->layers[layer].fmt)
+				if ((u32)(layer_info.src_surf.fmt) !=
+				    (u32)(disp->layers[layer].fmt))
 					vdsscomp_layer_disable(l);
 			} else
 				vdsscomp_layer_disable(l);
